fix: autoload when register offline table (#1872)
* fix: autoload when register offline table

* comment

* tests

* test offline
