+++ java/openmldb-batch/src/main/scala/com/_4paradigm/openmldb/batch/utils/HybridseUtil.scala
@@ -17,20 +17,26 @@
 package com._4paradigm.openmldb.batch.utils
 
 import java.util
-import com._4paradigm.hybridse.`type`.TypeOuterClass.{ColumnDef, Database, TableDef, Type}
-import com._4paradigm.hybridse.vm.{PhysicalLoadDataNode, PhysicalOpNode, PhysicalSelectIntoNode}
-import com._4paradigm.hybridse.node.{ConstNode, DataType => InnerDataType}
+
+import com._4paradigm.hybridse.`type`.TypeOuterClass.{ColumnDef, Database, TableDef}
+import com._4paradigm.hybridse.node.ConstNode
 import com._4paradigm.hybridse.sdk.UnsupportedHybridSeException
-import org.apache.spark.sql.types.{
-  BooleanType, DataType, DateType, DoubleType, FloatType, IntegerType, LongType,
-  ShortType, StringType, StructField, StructType, TimestampType
-}
-import org.apache.spark.sql.{DataFrame, Row}
+import com._4paradigm.hybridse.vm.{PhysicalLoadDataNode, PhysicalOpNode, PhysicalSelectIntoNode}
+import com._4paradigm.openmldb.proto
+import com._4paradigm.openmldb.proto.Common
+import org.apache.spark.sql.functions.{col, first}
+import org.apache.spark.sql.types.{DataType, LongType, StructField, StructType, TimestampType}
+import org.apache.spark.sql.{DataFrame, DataFrameReader, Row, SparkSession}
+import org.slf4j.LoggerFactory
+
 import scala.collection.JavaConverters.asScalaBufferConverter
+import scala.collection.convert.ImplicitConversions.`collection AsScalaIterable`
 import scala.collection.mutable
 
 
 object HybridseUtil {
+  private val logger = LoggerFactory.getLogger(this.getClass)
+
   def getOutputSchemaSlices(node: PhysicalOpNode, enableUnsafeRowOpt: Boolean): Array[StructType] = {
     if (enableUnsafeRowOpt) {
       // If enabling UnsafeRowOpt, return row with one slice
@@ -182,5 +188,98 @@ object HybridseUtil {
       deepCopy = Option(parseOption(getOptionFromNode(node, "deep_copy"), "true", getBoolOrDefault).toBoolean)
     }
     (format, options.toMap, mode, deepCopy)
+  }
+
+  // result 'readSchema' & 'tsCols' is only for csv format, may not be used
+  def extractOriginAndReadSchema(columns: util.List[Common.ColumnDesc]): (StructType, StructType, List[String]) = {
+    var oriSchema = new StructType
+    var readSchema = new StructType
+    val tsCols = mutable.ArrayBuffer[String]()
+    columns.foreach(col => {
+      var ty = col.getDataType
+      oriSchema = oriSchema.add(col.getName, SparkRowUtil.protoTypeToScalaType(ty), !col
+        .getNotNull)
+      if (ty.equals(proto.Type.DataType.kTimestamp)) {
+        tsCols += col.getName
+        // use string to parse ts column, to avoid getting null(parse wrong format), can't distinguish between the
+        // parsed null and the real `null`.
+        ty = proto.Type.DataType.kString
+      }
+      readSchema = readSchema.add(col.getName, SparkRowUtil.protoTypeToScalaType(ty), !col
+        .getNotNull)
+    }
+    )
+    (oriSchema, readSchema, tsCols.toList)
+  }
+
+  def parseLongTsCols(reader: DataFrameReader, readSchema: StructType, tsCols: List[String], file: String)
+  : List[String] = {
+    val longTsCols = mutable.ArrayBuffer[String]()
+    if (tsCols.nonEmpty) {
+      // normal timestamp format is TimestampType(Y-M-D H:M:S...)
+      // and we support one more timestamp format LongType(ms)
+      // read one row to auto detect the format, if int64, use LongType to read file, then convert it to TimestampType
+      // P.S. don't use inferSchema, cuz we just need to read the first non-null row, not all
+      val df = reader.schema(readSchema).load(file)
+      // check timestamp cols
+      for (col <- tsCols) {
+        val i = readSchema.fieldIndex(col)
+        var ty: DataType = LongType
+        try {
+          // value is string, try to parse to long
+          df.select(first(df.col(col), ignoreNulls = true)).first().getString(0).toLong
+          longTsCols.append(col)
+        } catch {
+          case e: Any =>
+            logger.debug(s"col '$col' parse long failed, use TimestampType to read", e)
+            ty = TimestampType
+        }
+
+        val newField = StructField(readSchema.fields(i).name, ty, readSchema.fields(i).nullable)
+        readSchema.fields(i) = newField
+      }
+    }
+    longTsCols.toList
+  }
+
+  // We want df with oriSchema, but if the file format is csv:
+  // 1. we support two format of timestamp
+  // 2. spark read may change the df schema to all nullable
+  // So we should fix it.
+  def autoLoad(spark: SparkSession, file: String, format: String, options: Map[String, String], columns: util
+  .List[Common.ColumnDesc]): DataFrame = {
+    val reader = spark.read.options(options)
+    val (oriSchema, readSchema, tsCols) = HybridseUtil.extractOriginAndReadSchema(columns)
+    if (format != "csv") {
+      return reader.schema(oriSchema).format(format).load(file)
+    }
+    // csv should auto detect the timestamp format
+
+    logger.info(s"set file format: $format")
+    reader.format(format)
+    // use string to read, then infer the format by the first non-null value of the ts column
+    val longTsCols = HybridseUtil.parseLongTsCols(reader, readSchema, tsCols, file)
+    logger.info(s"read schema: $readSchema, file $file")
+    var df = reader.schema(readSchema).load(file)
+    if (longTsCols.nonEmpty) {
+      // convert long type to timestamp type
+      for (tsCol <- longTsCols) {
+        df = df.withColumn(tsCol, (col(tsCol) / 1000).cast("timestamp"))
+      }
+    }
+
+    // if we read non-streaming files, the df schema fields will be set as all nullable.
+    // so we need to set it right
+    logger.info(s"after read schema: ${df.schema}")
+    if (!df.schema.equals(oriSchema)) {
+      df = df.sqlContext.createDataFrame(df.rdd, oriSchema)
+    }
+
+    require(df.schema == oriSchema, "df schema must == table schema")
+    if (logger.isDebugEnabled()) {
+      logger.debug("read dataframe count: {}", df.count())
+      df.show(10)
+    }
+    df
   }
 }
